refactor(parse): move decorators directory to lib
