+++ src/index.ts
@@ -1,2 +1,2 @@
-export * from './decorators';
+export * from '@lib/decorators';
 export * from '@lib/environment';
