+++ src/lib/decorators/parse-environment/types/parse-environment-options.type.ts
